Add LuckPerms integration to handle Discord linking
A new class, ServerLoginNetworkHandlerMixin, was added to handle user authenticity for Discord linking. The LuckPerms permissions library was added as a dependency to manage user permissions and help with their identification. Kotlin module were updated to include LuckPerms version. Also, an exclusive content repository was added to gradle for LuckPerms API lookup. Implementation of LuckPerms was also added to fabric and forge build files. Finally, the discordloom-common.mixins.json was also updated to include the mixin configuration for the newly introduced class.