Refactor package structure and enhance server login process
Refactored the package structure by moving ServerLoginNetworkHandlerMixin to a new server package. Added functionality to show a message and redirect to Discord for users attempting to join without the DiscordLoom mod. Created ClientLinkManager to manage the linking process and hold the URI string, and added a new mixin for handling disconnection screen. The changes are aimed at guiding users to install the necessary mod and link their accounts for a better gaming experience.