"Update integration method for Discord4j libraries in gradle files"
The previous implementation method for Discord4j libraries was changed from modImplementation to api in the common build.gradle and to shadowCommon in the fabric build.gradle file. The change was necessary to improve the exposure of these libraries to other modules and configurations to help with code cohesion and reusability.